requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Flask-WTF==1.2.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Werkzeug==3.0.1
WTForms==3.1.1
email-validator==2.1.0
gunicorn==21.2.0
pandas==2.1.4
openpyxl==3.1.2
groq>=0.12.0
Flask-Mail==0.9.1
Flask-Limiter==3.6.0
requests==2.31.0
scikit-learn==1.3.2
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
pypdf==4.0.1
twilio==8.11.0
weasyprint>=61.0
python-barcode[images]==0.16.1
reportlab==4.4.9
Pillow==10.1.0
